Fix manual exp manual dmd

+++ scripts/eval.py
@@ -193,7 +193,7 @@ def main():
             X_hat = model.rollout(K=K, C=C, x0=x0, steps=steps).cpu().numpy()
 
         else:
-            x0_torch = torch.tensor(x0, dtype=torch.float32)
+            x0_torch = torch.tensor(x0, dtype=torch.float64)
             X_hat = rollout_ae_model(
                 model,
                 x0=x0_torch,
@@ -239,7 +239,7 @@ def main():
     elif args.model == "manual_expansion_manual_dmd":
         X_hat = model.rollout(K=K, C=C, x0=x0, steps=steps).cpu().numpy()
     else:
-        x0_torch = torch.tensor(x0, dtype=torch.float32)
+        x0_torch = torch.tensor(x0, dtype=torch.float64)
         X_hat = rollout_ae_model(
             model,
             x0=x0_torch,
